frontend completed

+++ backend/src/controllers/userController.js
@@ -25,10 +25,17 @@ export const getUsers = async (req, res) => {
 // Delete a user
 export const deletUser = async (req, res) => {
   try {
-    const user  = await User.findByIdAndDelete(req.params.id)
+    const user  = await User.findById(req.params.id)
     if (!user) {
       return res.status(404).json({ message: "User not found" })
     }
+
+    if (user.role === 'admin') {
+      return res.status(403).json({ message: "Cannot delete admin user" })
+    }
+    
+    await User.findByIdAndDelete(req.params.id)
+
     res.status(200).json({ message: "User deleted successfully" })
   } catch (error) {
     console.error("Error deleting user: ", error);
